Add HTTPHandlers struct to prepare for 2 endpoints
