+++ go/v1_test.go
@@ -2,13 +2,11 @@ package zeroeventhub
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
-	"strconv"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -19,16 +17,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func mustMarshalJson(e any) json.RawMessage {
-	result, err := json.Marshal(e)
-	if err != nil {
-		panic(err)
-	}
-	return result
-}
-
 func createZehClientWithPartitionCount(server *httptest.Server, partitionCount int) Client {
-	return NewClient(fmt.Sprintf("%s/feed/v1", server.URL), partitionCount)
+	return NewClient(fmt.Sprintf("%s/testfeed", server.URL), partitionCount)
 }
 
 func createZehClient(server *httptest.Server) Client {
@@ -45,96 +35,8 @@ type TestZeroEventHubAPI struct {
 	partitions map[int][]TestEvent
 }
 
-func NewTestZeroEventHubAPI() *TestZeroEventHubAPI {
-	api := TestZeroEventHubAPI{partitions: map[int][]TestEvent{}}
-	partition0 := make([]TestEvent, 10000)
-	partition1 := make([]TestEvent, 10000)
-	for i := 0; i < 10000; i++ {
-		partition0[i] = TestEvent{
-			ID:      fmt.Sprintf("00000000-0000-0000-0000-%012x", i),
-			Version: 0,
-			Cursor:  i,
-		}
-		partition1[i] = TestEvent{
-			ID:      fmt.Sprintf("11111111-0000-0000-0000-%012x", i),
-			Version: 0,
-			Cursor:  i,
-		}
-	}
-	api.partitions[0] = partition0
-	api.partitions[1] = partition1
-	return &api
-}
-
-func (t TestZeroEventHubAPI) GetName() string {
-	return "TestZeroEventHubAPI"
-}
-
-func (t TestZeroEventHubAPI) GetPartitionCount() int {
-	return 2
-}
-
-func (t TestZeroEventHubAPI) FetchEvents(ctx context.Context, cursors []Cursor, pageSizeHint int, r EventReceiver, headers ...string) error {
-	if pageSizeHint == DefaultPageSize {
-		pageSizeHint = 100
-	}
-	for _, cursor := range cursors {
-		partition, ok := t.partitions[cursor.PartitionID]
-		if !ok {
-			return ErrPartitionDoesntExist
-		}
-		var err error
-		var lastProcessedCursor int
-		switch cursor.Cursor {
-		case FirstCursor:
-			lastProcessedCursor = -100
-		case LastCursor:
-			lastProcessedCursor = len(partition) - 2
-		// Mock responses: set the cursor to one of the following values to get a mocked response.
-		case cursorReturn500:
-			return err500
-		case cursorReturn504:
-			return err504
-		default:
-			lastProcessedCursor, err = strconv.Atoi(cursor.Cursor)
-			if err != nil {
-				return err
-			}
-		}
-		eventsProcessed := 0
-		h := make(map[string]string, 1)
-		for _, header := range headers {
-			if header == "content-type" {
-				h["content-type"] = "application/json"
-				break
-			}
-			if header == All {
-				h["content-type"] = "application/json"
-				h["foo"] = "bar"
-				break
-			}
-		}
-		for _, event := range partition {
-			if event.Cursor > lastProcessedCursor {
-				if err := r.Event(cursor.PartitionID, h, mustMarshalJson(partition[event.Cursor])); err != nil {
-					return err
-				}
-				if err := r.Checkpoint(cursor.PartitionID, fmt.Sprintf("%d", event.Cursor)); err != nil {
-					return err
-				}
-				lastProcessedCursor = event.Cursor
-				eventsProcessed++
-			}
-			if eventsProcessed == pageSizeHint {
-				break
-			}
-		}
-	}
-	return nil
-}
-
 func TestAPI(t *testing.T) {
-	server := httptest.NewServer(Handler("/feed/v1", nil, NewTestZeroEventHubAPI()))
+	server := Server(NewTestZeroEventHubAPI())
 	tests := []struct {
 		name string
 
@@ -258,7 +160,7 @@ func TestAPI(t *testing.T) {
 func BenchmarkFeed(b *testing.B) {
 	logger := logrus.New()
 	logger.SetOutput(io.Discard)
-	server := httptest.NewServer(Handler("/feed/v1", logger, NewTestZeroEventHubAPI()))
+	server := Server(NewTestZeroEventHubAPI())
 	client := createZehClient(server).WithLogger(logger)
 	var page EventPageSingleType[TestEvent]
 	err := client.FetchEvents(context.Background(), []Cursor{
@@ -292,7 +194,7 @@ func (l *loggingRoundTripper) RoundTrip(request *http.Request) (*http.Response, 
 }
 
 func TestJSON(t *testing.T) {
-	server := httptest.NewServer(Handler("/feed/v1", nil, NewTestZeroEventHubAPI()))
+	server := Server(NewTestZeroEventHubAPI())
 	loggingClient := server.Client()
 	loggingRoundTripper := loggingRoundTripper{actualRoundTripper: server.Client().Transport}
 	loggingClient.Transport = &loggingRoundTripper
@@ -389,7 +291,7 @@ func TestNewLines(t *testing.T) {
 }
 
 func TestRequestProcessor(t *testing.T) {
-	server := httptest.NewServer(Handler("/feed/v1", nil, NewTestZeroEventHubAPI()))
+	server := Server(NewTestZeroEventHubAPI())
 	loggingClient := server.Client()
 	loggingRoundTripper := loggingRoundTripper{actualRoundTripper: server.Client().Transport}
 	loggingClient.Transport = &loggingRoundTripper
@@ -405,7 +307,7 @@ func TestRequestProcessor(t *testing.T) {
 }
 
 func TestEnvelopeHeaders(t *testing.T) {
-	server := httptest.NewServer(Handler("/feed/v1", nil, NewTestZeroEventHubAPI()))
+	server := Server(NewTestZeroEventHubAPI())
 	client := createZehClient(server)
 	var page EventPageSingleType[TestEvent]
 	err := client.FetchEvents(context.Background(), []Cursor{{Cursor: LastCursor}}, DefaultPageSize, &page, "123")
@@ -443,33 +345,33 @@ func MockHandler(logger logrus.FieldLogger, api EventPublisher) http.Handler {
 	if logger == nil {
 		logger = logrus.StandardLogger()
 	}
-	router := mux.NewRouter()
-	router.Methods(http.MethodGet).
-		Path("/feed/v1").
-		HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-			query := request.URL.Query()
-			cursors, err := parseCursors(api.GetPartitionCount(), query)
-			if err != nil {
-				http.Error(writer, err.Error(), http.StatusBadRequest)
-				return
-			}
+	handlerFunc := func(writer http.ResponseWriter, request *http.Request) {
+		if request.URL.Path != "/testfeed" {
+			writer.WriteHeader(http.StatusNotFound)
+			return
+		}
 
-			serializer := NewNDJSONEventSerializer(writer)
-			err = api.FetchEvents(request.Context(), cursors, 10, serializer, All)
-			switch err {
-			case err500:
-				http.Error(writer, err.Error(), http.StatusInternalServerError)
-				return
-			case err504:
-				http.Error(writer, err.Error(), http.StatusGatewayTimeout)
-				return
-			default:
-				// Proceed
-			}
-		})
-	return http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		router.ServeHTTP(writer, request)
-	})
+		query := request.URL.Query()
+		cursors, err := parseCursors(api.GetPartitionCount(), query)
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusBadRequest)
+			return
+		}
+
+		serializer := NewNDJSONEventSerializer(writer)
+		err = api.FetchEvents(request.Context(), cursors, 10, serializer, All)
+		switch err {
+		case err500:
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+			return
+		case err504:
+			http.Error(writer, err.Error(), http.StatusGatewayTimeout)
+			return
+		default:
+			// Proceed
+		}
+	}
+	return http.HandlerFunc(handlerFunc)
 }
 
 func TestMockResponses(t *testing.T) {
